pyproject.toml
[project]
name = "llmgradingexperiments"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.18",
    "langchain-mistralai>=0.2.6",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.9.6",
]

[tool.ruff]
fix=true
unsafe-fixes=true
line-length = 120
lint.select = ["E", "F", "UP", "B", "SIM", "I"]
lint.ignore = ["COM812", "F401", "D203", "D211", "D213", "D100", "D104", "RUF001"]
